changes requested by hlib
