+++ rpc/coreclient_test.go
